requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
requests>=2.31.0
polars>=0.20.0
python-dotenv>=1.0.0
